x32: fix some bits headers where kernel and userspace long mismatch

+++ arch/x32/bits/msg.h
@@ -4,10 +4,10 @@ struct msqid_ds
 	time_t msg_stime;
 	time_t msg_rtime;
 	time_t msg_ctime;
-	unsigned long msg_cbytes;
+	unsigned long long msg_cbytes;
 	msgqnum_t msg_qnum;
 	msglen_t msg_qbytes;
 	pid_t msg_lspid;
 	pid_t msg_lrpid;
-	unsigned long __unused[2];
+	unsigned long long __unused[2];
 };
